+++ src/main/java/com/oracle/nosql/spring/data/core/mapping/NosqlCapacityMode.java
@@ -8,6 +8,8 @@ package com.oracle.nosql.spring.data.core.mapping;
 
 /**
  * Cloud only, selector for how capacity is set when creating new tables.
+ *
+ * @see oracle.nosql.driver.ops.TableLimits.CapacityMode
  */
 public enum NosqlCapacityMode {
     PROVISIONED,