+++ src/test/java/com/oracle/nosql/spring/data/test/id/SensorIdInstant.java
@@ -9,9 +9,7 @@ package com.oracle.nosql.spring.data.test.id;
 import java.time.Instant;
 
 import org.springframework.data.annotation.Id;
-import org.springframework.data.annotation.Persistent;
 
-@Persistent
 public class SensorIdInstant {
     String name;
     int temp;